Compute Sliced Wasserstein Barycenter

Computes the Wasserstein barycenter of multiple samples using the sliced Wasserstein distance. The barycenter is a representative distribution that minimizes the average transport cost to all input samples. This is used as the reference distribution for OT batch correction.

Usage

computeSlicedBarycenter(
  samplesList,
  nCellsPerSample = 2000L,
  nBarycenterCells = 5000L,
  nProjections = 100L,
  nIter = 10L,
  seed = 1994
)

Arguments

samplesList

List of data frames or matrices, one per sample. Each element contains cells (rows) x markers (columns).

nCellsPerSample

Integer; number of cells to subsample from each sample. Default is 2000.

nBarycenterCells

Integer; number of cells in the output barycenter. Default is 5000.

nProjections

Integer; number of random projections per iteration. Default is 100.

nIter

Integer; number of barycenter iterations. Default is 10.

seed

Integer; random seed. Default is 1994.

Value

Matrix of barycenter cells (nBarycenterCells x markers).

Examples

if (FALSE) { # \dontrun{
# Split decoded data by sample
decoded_df <- as.data.frame(decoded_data)
decoded_df$sample_id <- df$sample_id
samples_list <- split(decoded_df[, markers], decoded_df$sample_id)

# Compute barycenter
barycenter <- computeSlicedBarycenter(
  samplesList = samples_list,
  nCellsPerSample = 10000,
  nBarycenterCells = 5000,
  nProjections = 200,
  nIter = 20
)
} # }